Move already ported JR tests exercising `indexed-repeat` to dedicated suite
